Factor out common test utilities
